organizations update

+++ routes/web.php
@@ -32,15 +32,15 @@ Route::post('/logout', 'App\Http\Controllers\AuthController@logout')
 // ========== USER REQUESTS ===========
 // ====================================
 // user
-// Route::get('/user', 'App\Http\Controllers\ViewController@user')->name('user');
+ Route::get('/user', 'App\Http\Controllers\ViewController@user')->name('user');
 
 // Views
-// Route::get('/admin', 'App\Http\Controllers\ViewController@adminDashboard')->name('admin-dashboard');
-// Route::get('/admin/genre', 'App\Http\Controllers\ViewController@adminGenre')->name('admin.genre');
-// Route::get('/admin/project/rr', 'App\Http\Controllers\ViewController@adminRR')->name('admin-rr');
-// Route::get('/admin/project/ra', 'App\Http\Controllers\ViewController@adminRA')->name('admin-ra');
-// Route::get('/admin/project/rtp', 'App\Http\Controllers\ViewController@adminRTP')->name('admin-rtp');
-// Route::get('/admin/project/all', 'App\Http\Controllers\ViewController@adminAll')->name('admin-all');
+Route::get('/admin', 'App\Http\Controllers\ViewController@adminDashboard')->name('admin-dashboard');
+ Route::get('/admin/genre', 'App\Http\Controllers\ViewController@adminGenre')->name('admin.genre');
+ Route::get('/admin/project/rr', 'App\Http\Controllers\ViewController@adminRR')->name('admin-rr');
+ Route::get('/admin/project/ra', 'App\Http\Controllers\ViewController@adminRA')->name('admin-ra');
+Route::get('/admin/project/rtp', 'App\Http\Controllers\ViewController@adminRTP')->name('admin-rtp');
+ Route::get('/admin/project/all', 'App\Http\Controllers\ViewController@adminAll')->name('admin-all');
 
 // ====================================
 // ========== ADMIN REQUESTS ==========
@@ -69,15 +69,15 @@ Route::post('/logout', 'App\Http\Controllers\AuthController@logout')
 // USER MANAGEMENT
 // Route::get('/admin/user-management', 'App\Http\Controllers\Admin\UserManagementController@view')->name('user-management');
 
-// Route::group(['middleware' => 'AdminRoute'], function () {
-//     Route::get('/admin/pp', 'App\Http\Controllers\Admin\ProcessProfileController@view')->name('admin.pp');
-//     Route::post('/admin/pp/create', 'App\Http\Controllers\Admin\ProcessProfileController@create')->name('admin.pp.create');
-//     Route::get('/admin/tp', 'App\Http\Controllers\Admin\ThreatProfileController@view')->name('admin.tp');
-//     Route::post('/admin/tp/create', 'App\Http\Controllers\Admin\ThreatProfileController@create')->name('admin.tp.create');
-//     Route::get('/admin/vp', 'App\Http\Controllers\Admin\VulnProfileController@view')->name('admin.vp');
-//     Route::post('/admin/vp/create', 'App\Http\Controllers\Admin\VulnProfileController@create')->name('admin.vp.create');
-//     Route::get('/admin/genre', 'App\Http\Controllers\ViewController@adminGenre')->name('admin.genre');
-// });
+ Route::group(['middleware' => 'AdminRoute'], function () {
+     Route::get('/admin/pp', 'App\Http\Controllers\Admin\ProcessProfileController@view')->name('admin.pp');
+     Route::post('/admin/pp/create', 'App\Http\Controllers\Admin\ProcessProfileController@create')->name('admin.pp.create');
+    Route::get('/admin/tp', 'App\Http\Controllers\Admin\ThreatProfileController@view')->name('admin.tp');
+     Route::post('/admin/tp/create', 'App\Http\Controllers\Admin\ThreatProfileController@create')->name('admin.tp.create');
+     Route::get('/admin/vp', 'App\Http\Controllers\Admin\VulnProfileController@view')->name('admin.vp');
+     Route::post('/admin/vp/create', 'App\Http\Controllers\Admin\VulnProfileController@create')->name('admin.vp.create');
+    Route::get('/admin/genre', 'App\Http\Controllers\ViewController@adminGenre')->name('admin.genre');
+ });
 
 Route::middleware('user')->group(function () {
     Route::get('/user', 'App\Http\Controllers\ViewController@user')
@@ -109,17 +109,22 @@ Route::get('/admin/user-management', 'App\Http\Controllers\Admin\UserManagementC
 // ====================================
 // Standardized Routing Structure
 // ====================================
-
 use App\Http\Controllers\OrganizationController;
 
+
 Route::get('/admin/Organization', [OrganizationController::class, 'index']);
-Route::post('/organizations', [OrganizationController::class, 'store'])->name('organizations.store');
-Route::resource('organizations', OrganizationController::class);
-Route::put('/organizations/{organization}', [OrganizationController::class, 'update'])->name('organizations.update');
-Route::get('/organizations/{organization}/edit', [OrganizationController::class, 'edit'])->name('organizations.edit');
-Route::put('/organizations/{organization}', [OrganizationController::class, 'update'])->name('organizations.update');
-Route::get('/organizations/{organization}/edit', [OrganizationController::class, 'edit'])->name('organizations.edit');
-Route::delete('/organizations/{organization}', [OrganizationController::class, 'destroy'])->name('organizations.destroy');
+Route::post('/admin/organizations', [OrganizationController::class, 'store'])->name('organizations.store');
+Route::resource('/admin/organizations', OrganizationController::class);
+Route::put('/admin/organizations/{organization}', [OrganizationController::class, 'update'])->name('organizations.update');
+Route::get('/admin/organizations/{organization}/edit', [OrganizationController::class, 'edit'])->name('organizations.edit');
+Route::put('/admin/organizations/{organization}', [OrganizationController::class, 'update'])->name('organizations.update');
+Route::get('/admin/organizations/{organization}/edit', [OrganizationController::class, 'edit'])->name('organizations.edit');
+Route::delete('/admin/organizations/{organization}', [OrganizationController::class, 'destroy'])->name('organizations.destroy');
+Route::get('/admin/organizations/{organization}/edit', [OrganizationController::class, 'edit'])->name('organizations.edit');
+Route::resource('/admin/organizations', OrganizationController::class);
+Route::get('/admin/organizations/{organization}/edit', [OrganizationController::class, 'edit'])->name('organizations.edit');
+
+
 
 
 
@@ -149,3 +154,10 @@ use App\Http\Controllers\Admin\ProjectController;
 Route::get('/admin/projects', [ProjectController::class, 'view'])->name('admin.projects');
 Route::post('/admin/projects', [ProjectController::class, 'create'])->name('admin.projects.create');
 Route::patch('/admin/projects/{id}/update', [ProjectController::class, 'update'])->name('admin.projects.update');
+
+
+Route::group(['AdminRoute'], function () {
+    Route::get('/admin/tp', 'App\Http\Controllers\Admin\ThreatProfileController@view')->name('admin.tp');
+    Route::post('/admin/tp/create', 'App\Http\Controllers\Admin\ThreatProfileController@create')->name('admin.tp.create');
+});
+
